Explode plots.py into separate directory
BLD Add missing plots directory to installation.

ENH Add Rice Distributions (#1819)

Silenced Theano round warning

Changed all instances of tt.round to tround

Moved tround to pymc3.math

Fixed typo in tround

Argument fix for tround

Add Rice imports.

Revert "ENH Add Rice Distributions (#1819)"

This reverts commit 16996b5094274d062a01cfd953be6476f4592fa8.

Fix convergence detection bug in advi

BUG Stray import of Rice.

Make the pdf in the Lognormal docstring easier to read

Removed advi.ipynb and disaster_model.ipynb
