+++ pymc3/plots/__init__.py
@@ -1,0 +1,6 @@
+from .autocorrplot import autocorrplot
+from .compareplot import compare_plot
+from .forestplot import forestplot
+from .kdeplot import kdeplot, kde2plot
+from .posteriorplot import plot_posterior
+from .traceplot import traceplot
