+++ pymc3/plots/compareplot.py
@@ -1,0 +1,43 @@
+import matplotlib.pyplot as plt
+import numpy as np
+
+
+def compare_plot(comp_df, ax=None):
+    """
+    Model comparison summary plot in the style of the one used in the book
+    Statistical Rethinking by Richard McElreath.
+
+    Parameters
+    ----------
+
+    comp_df: DataFrame
+        The result of the pm.compare() function
+    ax : axes
+        Matplotlib axes. Defaults to None.
+
+    Returns
+    -------
+
+    ax : matplotlib axes
+
+    """
+    if ax is None:
+        _, ax = plt.subplots()
+
+    yticks_pos = np.linspace(0, -1, (comp_df.shape[0] * 2) - 1)
+    yticks_labels = np.repeat(comp_df.index, 2)[1:]
+
+    data = comp_df.values
+    min_ic = data[0, 0]
+
+    ax.errorbar(x=data[:, 0], y=yticks_pos[::2], xerr=data[:, 4], fmt='ko', mfc='None', mew=1)
+    ax.errorbar(x=data[1:, 0], y=yticks_pos[1::2], xerr=data[1:, 5], fmt='^', color='grey')
+
+    ax.plot(data[:, 0] - (2 * data[:, 1]), yticks_pos[::2], 'ko')
+    ax.axvline(min_ic, ls='--', color='grey')
+
+    ax.set_yticks(yticks_pos)
+    ax.set_yticklabels(yticks_labels)
+    ax.set_xlabel('Deviance')
+
+    return ax
