+++ setup.py
@@ -60,6 +60,7 @@ if __name__ == "__main__":
                     'pymc3.variational',
                     'pymc3.external',
                     'pymc3.gp',
+                    'pymc3.plots',
                     'docs',
                     '.',
           ],
